fix: 解决提取异常的情况
 对提取失败时关闭进度对话框

Log: 提取到U盘等移动设备，遇到移动设备在提取过程移除出现提取框不能自动消失；

Bug: https://pms.uniontech.com/zentao/bug-view-61861.html

+++ src/source/dialog/progressdialog.h
@@ -40,7 +40,7 @@ public:
 
     void setCurrentTask(const QString &file);
     void setCurrentFile(const QString &file);
-    void setProcess(int value);
+    void setProcess(double value);
     void setFinished();
     void clearprocess();
 
@@ -72,7 +72,7 @@ public slots:
 private:
     int m_defaultWidth = 380;
     int m_defaultHeight = 120;
-    int m_dPerent = 0; //保存上一次进度
+    double m_dPerent = 0.0; //保存上一次进度
 
     DLabel *m_tasklable;
     DLabel *m_filelable;
